petrol station asset
added petrol station. I have added it as singular meshes so that it can be more modular, but I have remade in engine the way I made it in the editor, so it can easily be copied. it's just in the test map and have kept all of the models, materials and textures neatly stored away