options menu
spent some time on formatting and had to redo the menu a couple of times to get it working properly. it is really finnicky so it has taken me a while. I've still got some stuff to polishj but basically the video menu is done, i've just got some errors with this json implementation.