YESSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSS!!!!!!!!
Finally loading from Json works!
The problem was that I kept the game settings as a monobehaviour, which can serialise as json but NOT deserialise.
This stumped me originally because I couldn't create a pointer to the gameSettings object.

The json utility could possibly be used to save and load games, which I may look into soon.

Game settings will save on quit and on load. My version of the game no longer uses the default unity settings and control interface and has a splash screen. The menu only needs to have audio settings change, this just hasn't been done yet because we don't have any proper audio. I'm pretty sure all I will have to do is create a mixer with 3 properties - music, sfx and master volume. So far it is done by searching for a tag, but this will be frustrating if we have to keep tagging stuff.
Apologies for this taking so long.