Menu prefab
Made a prefab of the menu for you. The only thing that's going to change now is the script for the audio so I shouldn't need to make another prefab.
The optionsManager script is attached to an empty object under the options tab, it should have everything attached all ready but it should be easy to work out - attach the ui objects to the settingsManager script.