menu audio settings
now have an audio mixer.
need to attach all sounds to the mixer and now the sliders will change
master
music
sfx
sounds

I spent ages wondering why it was loading with the correct slider location for audio but made no difference to the mixer - turns out (after about 3 hours of twiddling my thumbs trying to make heads or tails of it) that it only works a built version of the game. so...... a lot of wasted time but i'm glad it works.
Audio mixer gives us the oportunity to use reverbs, echos, distortion, lowpass, highpass filters and all that good stuff.