Main Game UI / Petrol Station Script
Messed about with the in game UI to make it look nicer. Still stuff to do, had to change stuff in the scene controller script. Not much, though.

Lines changed:
198
199
223

Also made the Petrol Station script which, upon entering the triggerbox you can now press enter and if you have five coins you can get a can of fuel. I need to add the pop up message so that players will know they can do that. Also need to balance out the use of petrol and the price / scarcity of it. We will have to discuss this together though.