Petrol Station
If you are in the car you now cannot buy petrol. If you buy petrol and you have a full tank you will get a can of fuel, otherwise you will get a full tank.
The menu now has a cursor, for some reason exiting to the menu from the game breaks the cursor, so now the cursor starts when the scene starts.