Safe Zone
Added a lot of models for use in levels

created the "safe zone" garage
in process of making a safe zone prefab
already done:
fences around safe zone
gate with opening doors
garage

to do: props, interactivity etc

Added a script that attaches to objects in the world which will make a sound when you crash into them such as a fence or a rock.

added fence hit sounds, metal gate hitting sounds

need to find some more "impact" sounds

created a new level to test out the free "EasyRoads3D" addon that can be used to make roads for the player to drive on. it's a bit -redacted- but it will work for what we need

currently on the hunt for some more environmental props and things to clutter the world to make it interesting

also DO NOT try to merge this commit to main as I haven't deleted all the irrelevant files yet