Fence destruction
Added a new script so fences "destruct" upon impact.
if the player is travelling more than 30km/h then the fence with break, otherwise it will stay. I have tried to make it a bit more efficient than it is currently, but all attempts so far have failed, but it is fairly usable at this point. I just need to tidy up some bits, add a script for opening the door at the safe zone garage (and sort of figure out what we're going to let the player do there) and some other things.
Destroyed fences will also despawn after 20 seconds to keep the game optimised. We could leave them in because it actually doesn't hurt the frame rate but... might aswell, just leaves crap in the way otherwise.
Also trying to add more control to the soundOnHit script by letting us control the volume of sounds when it hits an object. Thinking that an object should have a different sound level depending on the player speed when impacting with the object.
Made another fence model (Fence2)
Also deleting unused things from the stuff i've downloaded to keep the project size down.
Added fence hit impact sounds that I tweaked a bit in audacity.